requirements.txt
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
Babel==2.13.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.0.0
bleach-allowlist==1.0.3
boto3==1.34.162
botocore==1.34.162
Brotli==1.1.0
cachetools==4.2.4
cairocffi==1.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
croniter==2.0.7
cryptography==43.0.3
cssselect==1.2.0
cssselect2==0.7.0
cssutils==2.9.0
cycler==0.12.1
decorator==5.1.1
docopt==0.6.2
dropbox==11.36.2
email-reply-parser==0.5.12
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.13.4
filetype==1.2.0
fonttools==4.55.8
-e git+https://github.com/frappe/frappe.git@6f3baedb6ed1ae34924c56005f4c7771bfd3bd69#egg=frappe
fsspec==2025.2.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==1.34.1
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.66.0
gunicorn @ git+https://github.com/frappe/gunicorn@2f5b4923399d18c28e4221f6514792b5cc7d2c21
hiredis==2.2.3
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.28.1
idna==3.10
ipython==8.15.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
kiwisolver==1.4.8
ldap3==2.9.1
lxml==5.3.0
markdown2==2.4.13
markdownify==0.11.6
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
maxminddb==2.6.3
maxminddb-geolite2==2018.703
# Editable Git install with no remote (medical_imaging==0.0.1)
-e /Users/user/Codes/Personal/BE/medical-imaging-system/apps/medical_imaging
monotonic==1.6
mpmath==1.3.0
networkx==3.4.2
num2words==0.5.14
numpy==2.2.2
oauthlib==3.2.2
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
parso==0.8.4
passlib==1.7.4
pdfkit==1.0.0
pexpect==4.9.0
phonenumbers==8.13.13
pickleshare==0.7.5
pillow==10.3.0
ply==3.11
posthog==3.0.2
premailer==3.10.0
prompt_toolkit==3.0.50
protobuf==3.20.3
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydyf==0.10.0
Pygments==2.19.1
PyJWT==2.8.0
PyMySQL==1.1.1
pyOpenSSL==24.2.1
pyotp==2.8.0
pyparsing==3.2.1
pypdf==3.17.4
pyphen==0.17.2
PyPika==0.48.9
PyQRCode==1.2.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.2
rauth==0.7.3
redis==4.5.5
requests==2.32.3
requests-oauthlib==1.3.1
RestrictedPython==8.0
rq==1.15.1
rsa==4.9
s3transfer==0.10.4
safetensors==0.5.2
semantic-version==2.10.0
sentry-sdk==1.37.1
six==1.17.0
smmap==5.0.2
soupsieve==2.6
sql_metadata==2.11.0
sqlparse==0.5.3
stack-data==0.6.3
stone==3.3.8
sympy==1.13.1
tenacity==8.2.3
terminaltables==3.1.10
timm==1.0.14
tinycss2==1.1.1
tomli==2.0.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traceback-with-variables==2.0.4
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==3.0.1
urllib3==2.3.0
vobject==0.9.9
wcwidth==0.2.13
weasyprint==59.0
webencodings==0.5.1
Werkzeug==3.0.6
Whoosh==2.7.4
xlrd==2.0.1
zopfli==0.2.3.post1
zxcvbn==4.4.28
